fixing some bugs for sge scripts

+++ pymicmac/workflow/run_distributed_workflow_sge_cluster/run_distributed_workflow_sge_job.py
@@ -11,33 +11,44 @@ remoteExeDir = sys.argv[4]
 localOutDirAbsPath = sys.argv[5]
 
 # Read the command information from the XML and the provided index
-e = etree.parse(configFile).getroot()
+if not os.path.isfile(configFile):
+    raise Exception(configFile + " could not be found!")
+e = etree.parse(open(configFile)).getroot()
 c = e.xpath("//id[starts-with(text(),'" + commandIndex + "_')]/parent::*")[0]
 commandId = c.find("id").text.strip()
 command = c.find("command").text.strip()
 imagesListFile = c.find("images").text.strip()
 requiredElements= c.find("require").text.strip().split()
 outputElements = c.find("output").text.strip().split()
+print("CommandId: " + commandId)
+print("Command: " + command)
+print("Images list file: " + imagesListFile)
+print("Required: " + ' '.join(requiredElements))
+print("Output: " + ' '.join(outputElements))
 
-configFileAbsPath = dataDirAbsPath + '/' + configFile
+# Check image list file can be accesses
 imageListFileAbsPath = dataDirAbsPath + '/' + imagesListFile
-
+if not os.path.isfile(imageListFileAbsPath):
+    raise Exception(imageListFileAbsPath + " could not be found!")
 listImages = open(imageListFileAbsPath,'r').read().split('\n')
 
+# Create a local working directory using the specified remoteExeDir
 lwd = remoteExeDir + '/' + commandId
-
 shutil.rmtree(lwd, True)
 os.makedirs(lwd)
-
+# Change directory to be in the local working dir
 os.chdir(lwd)
 
+# Create a output directory in the shared folder to copy back the results
 commandLocalOutDirAbsPath = localOutDirAbsPath + '/' + commandId
+shutil.rmtree(commandLocalOutDirAbsPath, True)
 os.makedirs(commandLocalOutDirAbsPath)
 
+# Copy all the images from the data directory (in the shared folder) to the local one
 for image in listImages:
     if image != '':
         os.system('cp ' + dataDirAbsPath + '/' + image + ' .')
-
+# Copy other required files and folders in the data directory
 for requiredElement in requiredElements:
     requiredElementAbsPath =  dataDirAbsPath + '/' + requiredElement
     if os.path.isfile(requiredElementAbsPath):
@@ -47,12 +58,16 @@ for requiredElement in requiredElements:
     else:
         raise Exception(requiredElementAbsPath + " could not be found!")
 
-utils_execution.executeCommandMonitor(commandId, command, lwd, False)
+# Run the execution of the command (which includes cpu, mem and disk monitoring)
+utils_execution.executeCommandMonitor(commandId, command, lwd, True)
 
-os.system('cp ' + commandId + '.log ' + commandLocalOutDirAbsPath)
-os.system('cp ' + commandId + '.mon ' + commandLocalOutDirAbsPath)
-os.system('cp ' + commandId + '.mon.disk ' + commandLocalOutDirAbsPath)
-
+# Copy the monitor files back to the output dir in the shared folder
+for f in ( commandId + '.log ' , commandId + '.mon ', commandId + '.mon.disk '):
+    if os.path.isfile(f):
+        os.system('cp ' + f + ' ' + commandLocalOutDirAbsPath)
+    else:
+        raise Exception(f + " could not be found!")    
+# Copy other output files and folders back to the output dir in the shared folder
 for outputElement in outputElements:
     if os.path.isfile(outputElement):
         os.system('cp ' + outputElement + ' ' + commandLocalOutDirAbsPath)
@@ -61,4 +76,5 @@ for outputElement in outputElements:
     else:
         raise Exception(outputElement + " could not be found!")
 
+# Clean the local working dir
 os.system('rm -r ' + lwd)
